packages/mcp-server/pyproject.toml
[project]
name = "wayforth-mcp"
version = "0.1.0"
description = "MCP server for Wayforth — discover AI services from any MCP-compatible client"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]
